+++ src/main/java/com/qring/restaurant/application/res/RestaurantSearchResDTOV1.java
@@ -1,4 +1,4 @@
-package com.qring.restaurant.application.v1.res;
+package com.qring.restaurant.application.res;
 
 import com.qring.restaurant.domain.model.RestaurantEntity;
 import lombok.AllArgsConstructor;
