+++ src/main/java/com/qring/restaurant/application/service/RestaurantServiceV1.java
@@ -1,12 +1,12 @@
-package com.qring.restaurant.application.v1.service;
+package com.qring.restaurant.application.service;
 
 import com.qring.restaurant.application.global.exception.EntityNotFoundException;
 import com.qring.restaurant.application.global.exception.UnauthorizedAccessException;
-import com.qring.restaurant.application.v1.res.RestaurantGetByIdResDTOV1;
-import com.qring.restaurant.application.v1.res.RestaurantIdTableResDTOV1;
-import com.qring.restaurant.application.v1.res.RestaurantPostResDTOV1;
-import com.qring.restaurant.application.v1.res.RestaurantSearchResDTOV1;
-import com.qring.restaurant.application.v1.scheduler.OperationStatusScheduler;
+import com.qring.restaurant.application.res.RestaurantGetByIdResDTOV1;
+import com.qring.restaurant.application.res.RestaurantPostResDTOV1;
+import com.qring.restaurant.application.res.RestaurantIdTableResDTOV1;
+import com.qring.restaurant.application.res.RestaurantSearchResDTOV1;
+import com.qring.restaurant.application.scheduler.OperationStatusScheduler;
 import com.qring.restaurant.domain.model.CategoryEntity;
 import com.qring.restaurant.domain.model.OperatingHourEntity;
 import com.qring.restaurant.domain.model.RegionEntity;
@@ -17,8 +17,8 @@ import com.qring.restaurant.domain.repository.CategoryRepository;
 import com.qring.restaurant.domain.repository.RegionRepository;
 import com.qring.restaurant.domain.repository.RestaurantRepository;
 import com.qring.restaurant.infrastructure.util.PassportUtil;
-import com.qring.restaurant.presentation.v1.req.PostRestaurantReqDTOV1;
-import com.qring.restaurant.presentation.v1.req.PutRestaurantReqDTOV1;
+import com.qring.restaurant.presentation.req.PostRestaurantReqDTOV1;
+import com.qring.restaurant.presentation.req.PutRestaurantReqDTOV1;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
