+++ src/main/java/com/qring/restaurant/presentation/controller/RestaurantControllerV1.java
@@ -1,11 +1,15 @@
-package com.qring.restaurant.presentation.v1.controller;
+package com.qring.restaurant.presentation.controller;
 
 import com.qring.restaurant.application.global.dto.ResDTO;
+import com.qring.restaurant.application.res.RestaurantGetByIdResDTOV1;
+import com.qring.restaurant.application.res.RestaurantIdTableResDTOV1;
+import com.qring.restaurant.application.res.RestaurantPostResDTOV1;
+import com.qring.restaurant.application.res.RestaurantSearchResDTOV1;
 import com.qring.restaurant.application.v1.res.*;
-import com.qring.restaurant.application.v1.service.RestaurantServiceV1;
+import com.qring.restaurant.application.service.RestaurantServiceV1;
 import com.qring.restaurant.infrastructure.docs.RestaurantControllerSwagger;
-import com.qring.restaurant.presentation.v1.req.PostRestaurantReqDTOV1;
-import com.qring.restaurant.presentation.v1.req.PutRestaurantReqDTOV1;
+import com.qring.restaurant.presentation.req.PostRestaurantReqDTOV1;
+import com.qring.restaurant.presentation.req.PutRestaurantReqDTOV1;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
